Move internal functions to _internal packages
